# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/auth/signup/admin": {
            "post": {
                "description": "Sign up an admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Sign up an admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User Name",
                        "name": "user_name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v1/companies": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get companies",
                "tags": [
                    "Companies"
                ],
                "summary": "Get companies",
                "responses": {}
            }
        },
        "/v1/companies/create-company": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a company",
                "tags": [
                    "Companies"
                ],
                "summary": "Create a company",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company Code",
                        "name": "company_code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Company Name",
                        "name": "company_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Company Logo",
                        "name": "logo",
                        "in": "formData"
                    }
                ],
                "responses": {}
            }
        },
        "/v1/companies/delete-company": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a company",
                "tags": [
                    "Companies"
                ],
                "summary": "Delete a company",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "company_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v1/companies/edit-company": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Edit a company",
                "tags": [
                    "Companies"
                ],
                "summary": "Edit a company",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "company_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Company Code",
                        "name": "company_code",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Company Name",
                        "name": "company_name",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Company Logo",
                        "name": "logo",
                        "in": "formData"
                    }
                ],
                "responses": {}
            }
        },
        "/v1/companies/{companyId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get company detail",
                "tags": [
                    "Companies"
                ],
                "summary": "Get company detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "companyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}